+++ ios/app_tracking_transparency_plus/Package.swift
@@ -11,11 +11,15 @@ let package = Package(
     products: [
         .library(name: "app-tracking-transparency-plus", targets: ["app_tracking_transparency_plus"])
     ],
-    dependencies: [],
+    dependencies: [
+        .package(name: "FlutterFramework", path: "../FlutterFramework")
+    ],
     targets: [
         .target(
             name: "app_tracking_transparency_plus",
-            dependencies: [],
+            dependencies: [
+                .product(name: "FlutterFramework", package: "FlutterFramework")
+            ],
             resources: [
                 .process("PrivacyInfo.xcprivacy"),
             ]